requirements.txt
selenium
python-telegram-bot
PyVirtualDisplay
pyyaml
telethon